Sessions and authorization success on teh backend, tokens are generated for the sessions, about to work on connecting the front end
