+++ routes/index.js
@@ -14,49 +14,49 @@ router.get("/", (req, res)=>{
 router.use("/api", apiRoutes);
 
 //SignUp and Register Authentication Routes
-router.post(
-  '/signupPassport',
-  passport.authenticate('signup', { session: false }),
-  async (req, res, next) => {
-    res.json({
-      message: 'Signup successful',
-      user: req.user
-    });
-  }
-);
+// router.post(
+//   '/signupPassport',
+//   passport.authenticate('signup', { session: false }),
+//   async (req, res, next) => {
+//     res.json({
+//       message: 'Signup successful',
+//       user: req.user
+//     });
+//   }
+// );
 
-router.post(
-  '/loginPassport',
-  async (req, res, next) => {
-    passport.authenticate(
-      'login',
-      async (err, user, info) => {
-        try {
-          if (err || !user) {
-            const error = new Error('An error occurred.');
-            console.log(info.message)
-            return next(error);
-          }
+// router.post(
+//   '/loginPassport',
+//   async (req, res, next) => {
+//     passport.authenticate(
+//       'login',
+//       async (err, user, info) => {
+//         try {
+//           if (err || !user) {
+//             const error = new Error('An error occurred.');
+//             console.log(info.message)
+//             return next(error);
+//           }
 
-          req.login(
-            user,
-            { session: false },
-            async (error) => {
-              if (error) return next(error);
+//           req.login(
+//             user,
+//             { session: false },
+//             async (error) => {
+//               if (error) return next(error);
 
-              const body = { _id: user._id, email: user.email };
-              const token = jwt.sign({ user: body }, 'TOP_SECRET');
+//               const body = { _id: user._id, email: user.email };
+//               const token = jwt.sign({ user: body }, 'TOP_SECRET');
 
-              return res.json({ token });
-            }
-          );
-        } catch (error) {
-          return next(error);
-        }
-      }
-    )(req, res, next);
-  }
-);
+//               return res.json({ token });
+//             }
+//           );
+//         } catch (error) {
+//           return next(error);
+//         }
+//       }
+//     )(req, res, next);
+//   }
+// );
 
 
 // If no API routes are hit, send the React app
